:recycle: refactor: move package structure under subdomain
