+++ kok-api/src/main/java/com/kok/kokapi/config/jpa/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.kok.kokapi.config;
+package com.kok.kokapi.config.jpa;
 
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Configuration;
